reorganize config section
