+++ snuba/admin/static/runtime_config/styles.tsx
@@ -1,0 +1,15 @@
+import { COLORS } from "../theme";
+
+const containerStyle = {
+  width: 1200,
+  maxWidth: "100%",
+};
+
+const linkStyle = {
+  cursor: "pointer",
+  fontSize: 13,
+  color: COLORS.TEXT_LIGHTER,
+  textDecoration: "underline",
+};
+
+export { containerStyle, linkStyle };
