+++ snuba/admin/static/runtime_config/types.tsx
@@ -1,0 +1,9 @@
+import { ReactNode } from "react";
+
+type ConfigKey = string;
+type ConfigValue = string | number;
+type ConfigType = "string" | "int" | "float";
+
+type RowData = [ReactNode, ReactNode, ReactNode, ReactNode];
+
+export { ConfigKey, ConfigValue, ConfigType, RowData };
