--- snuba/admin/static/types.tsx
@@ -1,3 +1,0 @@
-type ConfigType = "string" | "int" | "float";
-
-export { ConfigType };
